detector_transaction_detail/requirements.txt
kafka-python
requests
lxml
